+++ examples/cc_ilb/variables.tf
@@ -187,6 +187,13 @@ variable "allow_global_access" {
   default     = false
 }
 
+variable "support_access_enabled" {
+  type        = bool
+  description = "Enable a specific outbound firewall rule for Cloud Connector to be able to establish connectivity for Zscaler support access. Default is true"
+  default     = true
+}
+
+
 # BYO (Bring-your-own) variables list
 
 variable "byo_vpc" {
@@ -333,5 +340,11 @@ variable "fw_cc_mgmt_ssh_ingress_name" {
 variable "fw_cc_service_default_name" {
   type        = string
   description = "The name of the compute firewall created on the user defined Cloud Connector Service VPC Network permitting workload traffic to be sent to Zscaler"
+  default     = null
+}
+
+variable "fw_cc_mgmt_zssupport_tunnel_name" {
+  type        = string
+  description = "The name of the compute firewall created on the user defined Cloud Connector Management VPC Network permitting CC to establish zssupport tunnel"
   default     = null
 }